% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankAccount.R
\docType{class}
\name{BankAccount-class}
\alias{BankAccount-class}
\alias{BankAccount}
\title{A Reference Class to represent a bank account.}
\description{
A Reference Class to represent a bank account.
}
\section{Fields}{

\describe{
\item{\code{balance}}{A length-one numeric vector.}
}}

\section{Methods}{

\describe{
\item{\code{withdraw(x)}}{Withdraw money from account. Allows overdrafts}
}}

